Homepage update
+Widgetlar ayrıldı.
+Kodlar sadeleştirildi.
