+++ android/app/build.gradle.kts
@@ -27,7 +27,7 @@ android {
         applicationId = "com.example.afet_acil_durum_app"
         // You can update the following values to match your application needs.
         // For more information, see: https://flutter.dev/to/review-gradle-config.
-        minSdk = flutter.minSdkVersion
+        minSdk = 23
         targetSdk = flutter.targetSdkVersion
         versionCode = flutter.versionCode
         versionName = flutter.versionName
