+++ lib/pages/emergency_contact.dart
@@ -4,7 +4,7 @@ import 'package:afet_acil_durum_app/pages/settings.dart';
 import 'package:afet_acil_durum_app/pages/homepage.dart';
 import 'package:afet_acil_durum_app/pages/notification_page.dart';
 import 'package:afet_acil_durum_app/pages/map.dart';
-import 'package:afet_acil_durum_app/services/connectivity_service.dart';
+import 'package:afet_acil_durum_app/services/connectivity/connectivity_service.dart';
 import 'package:afet_acil_durum_app/themes/theme_controller.dart';
 
 class EmergencyContact extends StatefulWidget {
