+++ lib/services/notifications/notification_service.dart
@@ -1,0 +1,64 @@
+import 'package:flutter_local_notifications/flutter_local_notifications.dart';
+
+class NotificationService {
+  final notificationsPlugin = FlutterLocalNotificationsPlugin();
+
+  bool _isInitialized = false;
+
+  bool get isInitialized => _isInitialized;
+
+  // Initialize
+  Future<void> initNotification() async {
+    if(_isInitialized) return; // prevent re-init
+    await notificationsPlugin
+        .resolvePlatformSpecificImplementation<AndroidFlutterLocalNotificationsPlugin>()
+        ?.requestNotificationsPermission();
+
+
+    //prepare android init settings
+    const initSettingsAndroid = AndroidInitializationSettings('@mipmap/ic_launcher');
+
+    //prepare ios init settings
+    const initSettingsIOS = DarwinInitializationSettings(
+        requestAlertPermission: true,
+        requestBadgePermission: true,
+        requestSoundPermission: true
+    );
+
+    //init settings
+    const initSettings = InitializationSettings(
+      android: initSettingsAndroid,
+      iOS: initSettingsIOS,
+    );
+
+    // finally, initialize the plugin
+    await notificationsPlugin.initialize(initSettings);
+  }
+
+
+  // Notification Detail Setup
+  NotificationDetails notificationDetails() {
+    return const NotificationDetails(
+      android: AndroidNotificationDetails(
+        'daily_channel_id',
+        'Daily Notifications',
+        channelDescription: 'Daily Notification Channel',
+        importance: Importance.max,
+        priority: Priority.high,
+        playSound: true,
+        enableVibration: true,
+      ),
+      iOS: DarwinNotificationDetails(),
+    );
+  }
+
+  // Show Notification
+  Future<void> showNotification({
+    int id = 0,
+    String? title,
+    String? body,
+  }) async {
+    return notificationsPlugin.show(id, title, body, notificationDetails());
+  }
+}
+
