+++ lib/widgets/header_widget.dart
@@ -1,0 +1,98 @@
+import 'package:flutter/material.dart';
+import 'package:afet_acil_durum_app/services/connectivity/connectivity_service.dart';
+
+class HeaderWidget extends StatelessWidget {
+  final VoidCallback onAuthorityTap;
+  final ConnectivityService connectivityService;
+
+  const HeaderWidget({
+    Key? key,
+    required this.onAuthorityTap,
+    required this.connectivityService,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: [
+        GestureDetector(
+          onTap: onAuthorityTap,
+          child: Container(
+            padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
+            decoration: BoxDecoration(
+              gradient: LinearGradient(
+                colors: [Colors.indigo.shade600, Colors.indigo.shade400],
+                begin: Alignment.topLeft,
+                end: Alignment.bottomRight,
+              ),
+              borderRadius: BorderRadius.circular(20),
+              boxShadow: [
+                BoxShadow(
+                  color: Colors.indigo.shade200.withOpacity(0.3),
+                  blurRadius: 4,
+                  offset: Offset(0, 2),
+                ),
+              ],
+            ),
+            child: Row(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                Icon(
+                  Icons.admin_panel_settings,
+                  color: Colors.white,
+                  size: 18,
+                ),
+                const SizedBox(width: 6),
+                Text(
+                  'Yetkili',
+                  style: TextStyle(
+                    color: Colors.white,
+                    fontWeight: FontWeight.w600,
+                    fontSize: 12,
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ),
+        StreamBuilder<BaglantiDurumu>(
+          stream: connectivityService.baglantiStream,
+          initialData: connectivityService.mevcutDurum,
+          builder: (context, snapshot) {
+            return Container(
+              padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
+              decoration: BoxDecoration(
+                color: connectivityService.baglantiRengi().withOpacity(0.1),
+                borderRadius: BorderRadius.circular(20),
+                border: Border.all(
+                  color: connectivityService.baglantiRengi(),
+                  width: 1.5,
+                ),
+              ),
+              child: Row(
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  Icon(
+                    connectivityService.baglantiIkonu(),
+                    color: connectivityService.baglantiRengi(),
+                    size: 18,
+                  ),
+                  const SizedBox(width: 6),
+                  Text(
+                    connectivityService.baglantiTipiMetni(),
+                    style: TextStyle(
+                      color: connectivityService.baglantiRengi(),
+                      fontWeight: FontWeight.w600,
+                      fontSize: 11,
+                    ),
+                  ),
+                ],
+              ),
+            );
+          },
+        ),
+      ],
+    );
+  }
+}
